v4.2.3: Stream resilience, battery optimization, fast startup, Wavyn Find fix, active UI button feedback, in-app updater, and smart recommendation algorithm

+++ app/src/main/kotlin/iad1tya/echo/music/recognition/ShazamModels.kt
@@ -8,7 +8,7 @@ data class ShazamRequestBody(
     @SerializedName("signature")
     val signature: Signature,
     @SerializedName("timestamp")
-    val timestamp: Int,
+    val timestamp: Long,
     @SerializedName("timezone")
     val timezone: String
 )
@@ -26,7 +26,7 @@ data class Signature(
     @SerializedName("samplems")
     val samplems: Int,
     @SerializedName("timestamp")
-    val timestamp: Int,
+    val timestamp: Long,
     @SerializedName("uri")
     val uri: String
 )
